[VA]: refactor, code migration.
