+++ VirtualApp/app/src/main/java/io/virtualapp/home/HomePresenterImpl.java
@@ -1,25 +1,11 @@
 package io.virtualapp.home;
 
 import android.app.Activity;
-import android.content.pm.PackageInfo;
 
-import com.lody.virtual.client.core.VirtualCore;
-import com.lody.virtual.os.VUserInfo;
-import com.lody.virtual.os.VUserManager;
-import com.lody.virtual.remote.InstallResult;
-import com.lody.virtual.remote.InstalledAppInfo;
-
-import java.io.IOException;
-
-import io.virtualapp.VApp;
 import io.virtualapp.VCommends;
-import io.virtualapp.abs.ui.VUiKit;
-import io.virtualapp.home.models.AppData;
 import io.virtualapp.home.models.AppInfoLite;
-import io.virtualapp.home.models.MultiplePackageAppData;
-import io.virtualapp.home.models.PackageAppData;
 import io.virtualapp.home.repo.AppRepository;
-import io.virtualapp.home.repo.PackageAppDataStorage;
+import io.virtualapp.sys.Installd;
 import jonathanfinerty.once.Once;
 
 /**
@@ -54,129 +40,6 @@ class HomePresenterImpl implements HomeContract.HomePresenter {
 
     @Override
     public void addApp(AppInfoLite info) {
-        class AddResult {
-            private PackageAppData appData;
-            private int userId;
-            private boolean justEnableHidden;
-            private boolean addedToLaucher;
-        }
-        AddResult addResult = new AddResult();
-        VUiKit.defer().when(() -> {
-            InstalledAppInfo installedAppInfo = VirtualCore.get().getInstalledAppInfo(info.packageName, 0);
-            addResult.justEnableHidden = installedAppInfo != null;
-            addResult.addedToLaucher = false;
-            if (info.disableMultiVersion) {
-                addResult.justEnableHidden = false;
-            }
-            if (addResult.justEnableHidden) {
-                int[] userIds = installedAppInfo.getInstalledUsers();
-                int nextUserId = userIds.length;
-                /*
-                  Input : userIds = {0, 1, 3}
-                  Output: nextUserId = 2
-                 */
-                for (int i = 0; i < userIds.length; i++) {
-                    if (userIds[i] != i) {
-                        nextUserId = i;
-                        break;
-                    }
-                }
-                addResult.userId = nextUserId;
-
-                if (VUserManager.get().getUserInfo(nextUserId) == null) {
-                    // user not exist, create it automatically.
-                    String nextUserName = "Space " + (nextUserId + 1);
-                    VUserInfo newUserInfo = VUserManager.get().createUser(nextUserName, VUserInfo.FLAG_ADMIN);
-                    if (newUserInfo == null) {
-                        throw new IllegalStateException();
-                    }
-                }
-                boolean success = VirtualCore.get().installPackageAsUser(nextUserId, info.packageName);
-                if (!success) {
-                    throw new IllegalStateException();
-                }
-            } else {
-                PackageInfo pkgInfo = null;
-                try {
-                    pkgInfo = VApp.getApp().getPackageManager().getPackageArchiveInfo(info.path, 0);
-                    pkgInfo.applicationInfo.sourceDir = info.path;
-                    pkgInfo.applicationInfo.publicSourceDir = info.path;
-                } catch (Exception e) {
-                }
-                if(pkgInfo != null) {
-                    PackageAppData data = PackageAppDataStorage.get().acquire(pkgInfo.applicationInfo);
-                    addResult.appData = data;
-                    data.isInstalling = true;
-                    data.isFirstOpen = false;
-                    mView.refreshLauncherItem(data);
-                    addResult.addedToLaucher = true;
-                }
-
-                InstallResult res = mRepo.addVirtualApp(info);
-                if (!res.isSuccess) {
-                    if (addResult.appData != null) {
-                        // mView.removeAppToLauncher(addResult.appData);
-                    }
-                    throw new IllegalStateException();
-                }
-            }
-        }).then((res) -> {
-            if (addResult.appData == null) {
-                addResult.appData = PackageAppDataStorage.get().acquire(info.packageName);
-                addResult.addedToLaucher = false;
-            }
-        }).done(res -> {
-            boolean multipleVersion = addResult.justEnableHidden && addResult.userId != 0;
-            if (!multipleVersion) {
-                PackageAppData data = addResult.appData;
-                data.isInstalling = false;
-                data.isLoading = true;
-                if (addResult.addedToLaucher)
-                    mView.refreshLauncherItem(data);
-                else
-                    mView.refreshLauncherItem(data);
-                handleOptApp(data, info.packageName, true);
-            } else {
-                MultiplePackageAppData data = new MultiplePackageAppData(addResult.appData, addResult.userId);
-                data.isInstalling = false;
-                data.isLoading = true;
-                if (addResult.addedToLaucher)
-                    mView.refreshLauncherItem(data);
-                else
-                    mView.refreshLauncherItem(data);
-                handleOptApp(data, info.packageName, false);
-            }
-        });
-    }
-
-
-    private void handleOptApp(AppData data, String packageName, boolean needOpt) {
-        VUiKit.defer().when(() -> {
-            long time = System.currentTimeMillis();
-            if (needOpt) {
-                try {
-                    VirtualCore.get().preOpt(packageName);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-            time = System.currentTimeMillis() - time;
-            if (time < 1500L) {
-                try {
-                    Thread.sleep(1500L - time);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
-        }).done((res) -> {
-            if (data instanceof PackageAppData) {
-                ((PackageAppData) data).isLoading = false;
-                ((PackageAppData) data).isFirstOpen = true;
-            } else if (data instanceof MultiplePackageAppData) {
-                ((MultiplePackageAppData) data).isLoading = false;
-                ((MultiplePackageAppData) data).isFirstOpen = true;
-            }
-            mView.refreshLauncherItem(data);
-        });
+        Installd.addApp(info, model -> mView.refreshLauncherItem(model));
     }
 }
